Fix remaining IDs (int -> long)

+++ app/src/main/java/com/o4x/musical/util/TagUtil.java
@@ -147,10 +147,10 @@ public class TagUtil {
     }
 
     public static class ArtworkInfo {
-        public final int albumId;
+        public final long albumId;
         public final Bitmap artwork;
 
-        public ArtworkInfo(int albumId, Bitmap artwork) {
+        public ArtworkInfo(long albumId, Bitmap artwork) {
             this.albumId = albumId;
             this.artwork = artwork;
         }
